Prevent failed ejabberdctl commands before restart
If the ejabberdctl commands are run before ejabberd is finished re/starting, then they'll complain about 'Error: command "register" not known.' and fail. While there's perhaps a way to listen for ejabberd to be ready since I've defined it as a Chef service, I couldn't find it, so now the LWRP waits a reasonable amount of time and then retries a few times before giving up.

(Note the retries and retry_delay resource attributes have to be in the LWRP, and not the part of the recipe that invokes it. Not sure why errors don't bubble up? Oh well, I'll ask in office hours I guess.)