Move nginx to vine_shared with optional app_server
The nginx.conf file is now in vine_shared, and expects the app server from vine_web, but doesn't require it, so it can serve only /supervisor URLs if we need it to for the vine-leaves box.

I also simplifed moved the SSL certs to here, and simplified how they were rendered.

Also, move the .bash_history commands to here from vine_web.