Use custom supervisor service instead of cookbook
The community cookbook was behaving weirdly with starts and reloads, and I think it was because it wasn't properly defining commands for all of the actions. I wasn't really using the auto-templatizing of conf files that the cookbook offerred, and it also didn't have the numprocs configuration settings I was using even if I *did* want to use that stuff, so I decided to get rid of it and just install the package myself.

Although looking again, the supervisor community cookbook has been recently updated, so maybe I'll re-integrate it, sigh.