Use the supervisor cookbook, not custom templates
And also not custom services. Note that I needed to fork the Opscode cookbook (http://community.opscode.com/cookbooks/supervisor) because it didn't have support for the web dashboard, which I added (https://github.com/lehrblogger/supervisor), and I'll probably submit a pull request for that.

There are corresponding commits in vine-web and vine-xmpp that use this cookbooks LWRP instead of their own custom templates.

Note that things might be in a broken state with this commit, because there are a lot of file and directory permission changes that I'm going to put in the next commit.

Also, UGH this was a pain to get working, with file permissions and etc. Note that if I try to run supervisor as a non-root user, it starts complaining about the unix_http_server's .sock file. There were other headaches to that I don't really understand well enough to describe here.