Start MySQL before Supervisor and gunicorn
I was getting "OperationalError: (OperationalError) (2006, 'MySQL server has gone away')" errors from Gunicorn after re-provisioning. I *think* this is because MySQL was being restarted, and fresh database connections were being created, after Gunicorn had already been started by Supervisor. So, it's connections were suddenly inexplicably gone, and it was failing to recreate them.

Hopefully this fixes it, but reproducing this bug was somewhat sporadic, and required I re-provision to reset the state, so I'm going to move on for now...