WIP, but now it reads the encrypted dev data bag
Note that the typo in upload_all.sh took FOREVER to figure out. But now I understand the Vagrantfile a bit better. Vagrant will upload your secret_key automatically for you here:
	https://github.com/mitchellh/vagrant/blob/master/plugins/provisioners/chef/provisioner/chef_client.rb#L72-L76
And that functionality was added in this commit here:
	https://github.com/mitchellh/vagrant/pull/398/files
And the default is used here:
	https://github.com/opscode/chef/blob/master/lib/chef/encrypted_data_bag_item.rb#L300

So, what happens is that Vagrant will set the default location to /tmp/encrypted_data_bag_secret when you don't specify it in the host machine's Vagrantfile, and that location then gets written to the Vagrant Chef client's knife.rb filei in the template in the commit, which then gets used as the default for reading the data bag, instead of /etc/chef/encrypted_data_bag_secret. Deep breath. The #chef IRC channel is great.