Move MySQL and ejabberd dump loading to new recipe
So that rather than have a flag I need to set for, I can just incude the recipe or not depending on if I want to run the dump import.

NOTE that this commit is going to break without the next one (or more), which has a bunch of node attribute and databag refactoring. I was doing a bunch of different things at once.