Reorganize and rename roles for EC2
It was confusing having a "python" community cookbook and a "python" role, so the "python" role is now the "supervised" role, which has the recipies for supervising python processes, with nginx and supervisor for managing them, and possibly other stuff. Now vagrant doesn't use the roles, but the individual recipies, to prevent duplication. This also allows us to run the vine_shared::load_dumps recipe in the right place, which was going to be tricky to do with the other organization.