Move AWS/Chef from personal/Vine to Dashdash acct
I want to move all of the Dashdash stuff off of my personal AWS account and onto a new, separate AWS account. This was inspired by the desire to host my personal blog as a static site on S3 and keep expenses separate, but this was probably something I'll need to do eventually anyway, and it's only going to get harder, so now seemed like a good time. Unfortunately, Amazon does not make this easy: https://forums.aws.amazon.com/thread.jspa?threadID=24308

So, I set about moving stuff over. Things are currently in a half-finished state. EC2, Route53, SQS, and IAM are on the new account, but RDS, SNS, CloudWatch, stage EC2 instance are still on the old account. I also set up a new Hosted Chef Organization, since I wanted to have both machines updating at once. I'll delete the old one when it's all moved over.

This was all a bit of a pain, since the Chef recipes kept failing and it was slow to retry them, but the tweaks in these last few commits got everything to work in prod, with the new instances talking to the old database (probably with more lag than usual, but it's temporary).
	2a5196e6f000e054abe4bf4f8bb1c7c4185227d1
	4433c731cb9e671575c9a5394ece629cf1f9ea94
	55e00558b51e5c4ad161e8c97620632662344be1
(Although, grr, cpan sometimes inexplicably fails to install those libraries.)