Consolidate nodes, get Chef running, patch openssl
To save money going forward I'm going to try running everything on a single EC2 machine rather than three separate machines. This is not thoroughly tested, but seems to work fine.

I updated some things for the Heartbleed bug in openssl, and it was a bit tricky to get Chef to run again, which was frustrating.

In the next commit I'll upload new encrypted data bags with possibly-compromised keys and passwords.