Update prod data bag and templates with renewed SSL certificate
OOF that was painful. Somethign in the new Chef Development Kit seems to be unhappy with my data bag format, so I was getting errors like this:
	Chef::Exceptions::ValidationFailed: Option data_bag's value {"encrypted_data"=>"PffgOkpIpdoEJO8khrUOUQwqv2/vqrtzOf1U/z/a5xD4KqSH2/CkD1zHndzW\nwJL1\n", "iv"=>"d/kiiPRQWQoKBTU5WF8NPw==\n", "version"=>1, "cipher"=>"aes-256-cbc"} does not match regular expression /^[\-[:alnum:]_]+$/
(Output from http://stackoverflow.com/questions/35616044/chefexceptionsvalidationfailed-error-during-encrypteddatabagitem-load-due-to, not me.) I tried all of the different versions and nothing worked. Eventually I noticed that there was the old Chef Client still available at https://downloads.chef.io/chef-client/mac/, so I downloaded 10.30.4 (the closest to 10.32, which I am using) and tried that. I then got a different error:
	ERROR: JSON::ParserError: Unsupported `json_class` type '{"encrypted_data"=>"JW+QNVeXN5yPLgyqwSmqsA64MFjb0dI0FXgZ5p5eb1zdsu3UQtW1jGSQspcx\nLJDT\n", "iv"=>"jo/JlgjnwNFI5Re8g1QRbw==\n", "version"=>1, "cipher"=>"aes-256-cbc"}'
(Output from me, this time.) Finally I realized I could try overwriting the config on the server with the one on the client, and then after that I was able to edit it as before.

Note that I'm not sure why I didn't need to update the intermediate and root SSL certificates last year – I'm guessing I just didn't notice them, and things magically worked anyway. Fingers crossed that this works, since `sudo chef-client` now fails on my EC2 intance, so I'm hoping the hourly one does it's thing, and that I never have to do this again...